Initial inclusion of `libpng15{-emul32}` package
This is being provided purely for binary compatibility with older games
or apps that might explicitly use libpng15 ABI, as some also use libpng12,
and we're moving our libpng to the 16 ABI now.

Per other `binary.compat` packages, it is now forbidden to link directly
against `libpng15.so` in Solus, so there are no provided `-devel` packages
here.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>